사소한 리팩토링

+++ project/src/main/java/com/umbum/pos/repository/PaymentRepo.java
@@ -7,6 +7,6 @@ import com.umbum.pos.model.Payment;
 
 public interface PaymentRepo {
     public Payment read();
-    public long create(Payment payment);
-    public long createAll(List<Payment> paymentList);
+    public int create(Payment payment);
+    public int createAll(List<Payment> paymentList);
 }
