+++ project/src/main/java/com/umbum/pos/repository/SalesRepo.java
@@ -4,5 +4,5 @@ import com.umbum.pos.model.Sales;
 
 public interface SalesRepo {
     public Sales read();
-    public long create(Sales sales);
+    public int create(Sales sales);
 }
